refactoring math library
